# api/openapi.yaml
openapi: 3.1.0
info:
  title: cyber-databrew API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - GraceToken: []
components:
  securitySchemes:
    GraceToken:
      type: apiKey
      in: header
      name: X-Grace-Token
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        details:
          type: object
          additionalProperties: true
    AssetID:
      type: string
      minLength: 8
      maxLength: 8
      pattern: '^[0-9A-Za-z]{8}$'
      description: Primary key for assets (8 alphanumeric ASCII characters).
...
    SearchReindexJobListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SearchReindexJob' }
...
    QueryRequest:
      type: object
      required: [scope]
      properties:
        schema_version: { type: string, default: v1 }
        mode:
          type: string
          enum: [structured, keyword, semantic, similar]
        scope: { $ref: '#/components/schemas/QueryScope' }
        select: { $ref: '#/components/schemas/QuerySelect' }
        where: { $ref: '#/components/schemas/QueryExpr' }
        sort:
          type: array
          items: { $ref: '#/components/schemas/QuerySort' }
        page: { $ref: '#/components/schemas/QueryPage' }
        facets:
          type: array
          items: { $ref: '#/components/schemas/QueryFacet' }
        debug: { $ref: '#/components/schemas/QueryDebug' }
...
    QueryValidateResponse:
      type: object
      properties:
        valid: { type: boolean }
        normalized_query: { $ref: '#/components/schemas/QueryRequest' }
        warnings:
          type: array
          items: { type: string }
        field_capabilities:
          type: array
          items: { $ref: '#/components/schemas/FieldCapabilityBrief' }
        debug_plan: { $ref: '#/components/schemas/DebugPlan' }
...
paths:
  /api/v1/queries/validate:
    post:
      summary: Validate Query IR request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryRequest' }
      responses:
        '200':
          description: query is valid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryValidateResponse' }
        '400':
          description: syntax/shape error (INVALID_ARGUMENT)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: semantic/planning error (UNSUPPORTED_FIELD/UNSUPPORTED_OPERATOR/UNPLANNABLE_QUERY)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/queries/run:
    post:
      summary: Execute Query IR request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryRequest' }
      responses:
        '200':
          description: query result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueryRunResponse' }
        '400':
          description: syntax/shape error (INVALID_ARGUMENT)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: semantic/planning error (UNSUPPORTED_FIELD/UNSUPPORTED_OPERATOR/UNPLANNABLE_QUERY)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/saved-queries:
    get:
      summary: List saved queries
      responses:
        '200':
          description: saved queries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SavedQuery' }
    post:
      summary: Create saved query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, query_ir_json]
              properties:
                name: { type: string }
                description: { type: string }
                owner: { type: string }
                query_ir_json: { $ref: '#/components/schemas/QueryRequest' }
      responses:
        '201':
          description: created saved query
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedQuery' }
  /api/v1/saved-queries/{id}:
    get:
      summary: Get saved query
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: saved query
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedQuery' }
    patch:
      summary: Update saved query
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, query_ir_json]
              properties:
                name: { type: string }
                description: { type: string }
                owner: { type: string }
                query_ir_json: { $ref: '#/components/schemas/QueryRequest' }
      responses:
        '200':
          description: updated saved query
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedQuery' }
    delete:
      summary: Delete saved query
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: deleted
  /api/v1/internal/assets/{id}:
    delete:
      summary: Hard delete a single asset and all of its child rows
      description: >
        Privileged maintenance endpoint guarded by `X-Grace-Token`. Removes the
        asset row along with referencing rows in `asset_metrics`,
        `asset_eval_results`, `asset_algo_events`, `asset_algo_latest`,
        `asset_tags`, `actions`, `delivery_items`, `asset_relations` and
        `asset_events`. Returns `404` when the asset is not found. The
        associated `mcap_files` row is left in place — use the batch endpoint
        with `include_mcap_files=true` when you also want to drop mcap files.
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/AssetID' }
      responses:
        '200':
          description: deleted
          content:
            application/json:
              schema:
                type: object
                required: [asset_id, deleted]
                properties:
                  asset_id: { $ref: '#/components/schemas/AssetID' }
                  deleted: { $ref: '#/components/schemas/PurgeCounts' }
        '400':
          description: invalid asset id
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: asset not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/internal/assets:batch_delete:
    post:
      summary: Hard delete a batch of assets (and optionally their mcap_files)
      description: >
        Privileged maintenance endpoint guarded by `X-Grace-Token`. Exactly
        one of `asset_ids` or `import_batch` must be set. Deletions run in
        chunks of `chunk_size` ids (default `1000`), each chunk inside its
        own PostgreSQL transaction so progress survives partial failures and
        the CDC consumer can ingest deletes incrementally into Elasticsearch.
        Use `dry_run=true` to preview the per-table counts that would be
        deleted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeleteRequest' }
      responses:
        '200':
          description: counts of deleted (or to-be-deleted, when dry_run) rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchPurgeResponse' }
        '400':
          description: invalid request body, or neither/both of asset_ids and import_batch supplied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: persistence error (for example an mcap_file is still referenced by an asset that was not part of the purge)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/reindex:
    post:
      summary: Rebuild Elasticsearch `assets` index from PostgreSQL
      description: >
        Scans all non-deleted assets and bulk-indexes documents. Does not modify
        `asset_events` rows. Uses the normal `X-Grace-Token` auth path for now
        and requires reachable ES. When the `assets` index does not exist yet,
        the initial delete-by-query step treats HTTP 404 as zero deletions so the
        bulk index can proceed.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dry_run:
                  type: boolean
                  default: false
                page_size:
                  type: integer
                  maximum: 500
                  default: 200
      responses:
        '200':
          description: Summary counters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexResponse' }
        '401':
          description: invalid or missing admin token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Elasticsearch not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/reindex-jobs:
    post:
      summary: Create asynchronous Elasticsearch reindex job
      description: >
        Creates a background PG->ES reindex job. Supports `dry_run` preview,
        stop, and resume from checkpoint. Returns the created job immediately.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dry_run:
                  type: boolean
                  default: false
                page_size:
                  type: integer
                  maximum: 500
                  default: 200
      responses:
        '202':
          description: job accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexJob' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Elasticsearch or job repository not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: List recent asynchronous reindex jobs
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: recent jobs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexJobListResponse' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/reindex-jobs/{id}:
    get:
      summary: Get one asynchronous reindex job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: job detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexJob' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: job not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/reindex-jobs/{id}/stop:
    post:
      summary: Request stop for asynchronous reindex job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: stop requested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexJob' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: job not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/reindex-jobs/{id}/resume:
    post:
      summary: Resume paused/failed asynchronous reindex job from checkpoint
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: job resumed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchReindexJob' }
        '400':
          description: job is not resumable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: job not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/admin/search/outbox-stats:
    get:
      summary: Outbox publish_state counts and DLQ archive size (read-only)
      description: >
        Returns `SELECT publish_state, COUNT(*) FROM asset_events GROUP BY publish_state`
        plus `SELECT COUNT(*) FROM outbox_dlq`. Uses `X-Grace-Token` like other `/api/v1/*` routes.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchOutboxStatsResponse' }
        '401':
          description: invalid or missing token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Postgres asset_events repository not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/report:
    get:
      summary: Get local lakehouse MVP report
      description: Returns the latest Iceberg query report generated by `make iceberg-mvp` or `make iceberg-mvp-host`.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseReport' }
        '404':
          description: report not generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/lakehouse/tables:
    get:
      summary: List Iceberg MVP table row counts via BigQuery
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseTableListResponse' }
        '503':
          description: bigquery disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/overview:
    get:
      summary: Get lakehouse dashboard overview metrics
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseOverviewResponse' }
        '503':
          description: bigquery disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/event-daily:
    get:
      summary: Get recent daily event counts from gold_asset_event_daily
      parameters:
        - in: query
          name: days
          schema: { type: integer, minimum: 1, default: 30 }
          description: Number of recent days to include.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseEventDailyResponse' }
        '400':
          description: invalid days parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: bigquery disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/event-type-share:
    get:
      summary: Get event_type share for a selected date
      parameters:
        - in: query
          name: date
          schema: { type: string, default: latest, example: '2026-05-12' }
          description: Selected date in YYYY-MM-DD, or latest.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseEventTypeShareResponse' }
        '400':
          description: invalid date parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: bigquery disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/asset-growth:
    get:
      summary: Get daily new assets and running cumulative total from gold (asset_created)
      parameters:
        - in: query
          name: days
          schema: { type: integer, minimum: 1, maximum: 180, default: 30 }
          description: Lookback window in days (capped at 180).
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseAssetGrowthResponse' }
        '400':
          description: invalid days parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: bigquery disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/lakehouse/sync-status:
    get:
      summary: Get realtime or reconciled PG vs Iceberg sync status
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncStatusResponse' }
        '503':
          description: postgres unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/lakehouse/sync-progress:
    get:
      summary: PG→Bronze watermark-based sync progress (O(1) queries)
      description: |
        Returns 4 watermark fields for the PG → Iceberg Bronze pipeline (Cloud
        Run Job `bronze-incremental`). All values are cheap to compute even at
        50B+ rows: `outbox_published_max_seq` is a PG index lookup,
        `bronze_max_event_seq` and `bronze_last_ingested_at` come from Iceberg
        metadata via BigQuery without scanning parquet. Also sets Prometheus
        gauges `lakehouse_bronze_max_event_seq`, `lakehouse_bronze_lag_events`,
        `lakehouse_bronze_last_ingested_unix_seconds`.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BronzeSyncProgressResponse' }
        '503':
          description: bigquery or postgres unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/lakehouse/quality-distribution:
    get:
      summary: Query asset quality distribution via Lakehouse Silver current-state table
      parameters:
        - in: query
          name: window
          schema: { type: string, enum: [7d, 30d, 60d, 90d], default: 30d }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseItemsResponse' }
  /api/v1/lakehouse/customer-replay:
    get:
      summary: Query customer delivery replay manifest via BigQuery
      parameters:
        - in: query
          name: customer_id
          schema: { type: string, default: 'urn:grace:customer:A' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LakehouseItemsResponse' }
  /api/v1/assets:
    post:
      summary: Create asset
      responses:
        '201': { description: created }
        '422':
          description: invalid domain state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/assets/{id}:
    get:
      summary: Get asset
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
        '404':
          description: not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    patch:
      summary: Update asset
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
        '404':
          description: not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: concurrent conflict (asset was modified between GET and PATCH)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      summary: Soft-delete asset
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
...